+++ site/src/content/docs/history.md
@@ -6,6 +6,12 @@ section: meta
 Previous releases are available on Github as dll files. There is no support for older releases. If you are looking for
 the latest version, go to the [Home Page](./) instead.
 
+## Version 8.25
+* [Minor] The texmod module now works with pre-loaded instances of gMod
+
+## Version 8.24
+* [Fix] Fixed an imgui bug causing some elements to be duplicated
+
 ## Version 8.23
 * [Fix] Various bug fixes
 * [New] Added texmod module to load texmods without needing to pre-launch umod/texmod/gmod